4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -5,16 +5,6 @@ const bodyParser = require('body-parser')
 blogsRouter.use(bodyParser.json())
 const jwt = require('jsonwebtoken')
 
-/*
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-      return authorization.substring(7)
-    }
-    return null
-}
-*/
-
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
     .find({}).populate('user', { username: 1, name: 1 })
@@ -25,7 +15,6 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
 
-    //const token = getTokenFrom(request)
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
     if (!request.token || !decodedToken.id) {
@@ -89,9 +78,25 @@ blogsRouter.put('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-    try {
-      await Blog.findByIdAndRemove(request.params.id)
-      response.status(204).end()
+    try{
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+        if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+    
+        const blog = await Blog.findById(request.params.id)
+
+        if(!blog){
+            return response.status(404).end()
+        }
+
+        if ( blog.user.toString() === decodedToken.id.toString() ){
+            await Blog.findByIdAndRemove(request.params.id)
+
+            response.status(204).end()
+        }
     } catch (exception) {
       next(exception)
     }
